+++ change-log.md
@@ -1,5 +1,17 @@
 # Change log
 
+## 3.0.5
+
+* FIXED[#69](https://github.com/mnapoli/PHP-DI/issues/69): Definition conflict when setting a value for a class name
+
+## 3.0.4
+
+* FIXED[#70](https://github.com/mnapoli/PHP-DI/issues/70): YamlDefinitionFileLoader crashes if YAML file is empty
+
+## 3.0.3
+
+* Fixed over-restrictive dependencies in composer.json
+
 ## 3.0.2
 
 * [#64](https://github.com/mnapoli/PHP-DI/issues/64): Non PHP-DI exceptions are not captured-rethrown anymore when injecting dependencies (cleaner stack trace)